requirements.txt
agno
python-dotenv
pydantic
pytest-asyncio
google-genai    
